requirements.txt
discord.py~=1.6.0
requests~=2.25.1
lxml~=4.6.2
pynacl~=1.4.0
